Added surface mode CCR summary.
This custom view gives a one page overview of the relevant settings for
CCR diving:
- sensor mode;
- setpoints / change depths;
- diluents / change depths;
- scrubber time left (if timer is used).
This allows CCR divers to quickly and easily confirm correct settings
before starting a dive.
The view is only available at in surface mode, and only if the dive mode
is CCR.

Additionally, the OC gaslist view has been added to the views available
in surface mode - this allows easy confirmation that correct OC gases /
bailout gases in CCR mode are configured.